major refactor
target interface now takes in a context object which should be used to
execute shell comands and log messages

old context directory is now called builder which seems like a more apt
name.

target interface is moved out to sevki.org/build/build so now we can do

       build.NewContext()

to pass in to the build function.