builder: fix: coordinator builds multiple times
While this wasn't a show stopper, it sure slowed it down. Since I hadn't
done a lot of real world tests, this never came up but basically if
while target a was building another target a was sent to the job queue
it would also start building. Because build insists on hashes are as
deterministic as target outputs this never affected the build process
which is why I didn't realize it.

Changes:
- When a node is created, number of it's dependencies is added to it's
workgroup
- Nodes keep a list of it's Parents and Children, parents for notifying
that one of their dependencies are done, and children to install their
artifacts.
- Nodes now hold a Lock so other workers can't build the same node at
the same time