postprocessor: added post processor
Dependencies and fields with build tags that are equal to 'path' are
processed.

Dependency postprocessing is pretty straight forward, dependencies that
start with ':' are concatted with the package string.

Path postprocessing is a little bit more complicated, files that start
with '//' are absoluted, files that aren't will be absoluted by joining
the project path, package path and the file path. There is a caveat for
file paths, which is; if the file path that is being absoluted is a file
that's installed by a dependency it will be not be absoluted.

for instance if a exports

    { "x.c": "b_output" }

and b depends on a at the same time b has "b_output/x.c" in it's source
list b_output/x.c will not be absaluted.