processor: processor now emits build.Targets trough a chan
this now brings us to the point where we read only as many bytes
as we need.

For instance if the build tool is pointed to a file that has a logic
errors in it, ie; duplicate load statements, targets that have
duplicate dependencies, build will quit before reading the entire
file.