targets/cc: add strip, add alwayslink
if a cc_binary has alwayslink set true it will look for libfoo.a
in its buildspace.

when dependencies are kernel libraries (starts with klib) build will
act as if alwayslink is on.

Essentially there are 3 types of linking,

	fully static:
where all symbols are linked by passing their .a files to the
compiler
	mostly static:
linked against .a files in lib search paths
	shared:
linked against .so files

alwayslink option for cc_binaries mean looking for .a files in the
buildspace