project: copy to root
with this change bldy copy things to root of the project instead
of build out. this was possible before with by manually setting
the BUILD_OUT envinroment variable but now it can either
be triggered by

	bldy -r kernel

-r flag now stands for root.
by setting COPYTOROOT=true in either bldy.cfg in the root of the
project or as an envinroment variable.

closes #57

Signed-off-by: Sevki <s@sevki.org>