project: remove the "should not be used with watch" statement
It is a generic problem that every build tool and compiler has
If you are watching a directory for changes to recompile a project
and your compiler writes to that directory the changes will cause
an infinite loop. But that certainly applies for any dir watcher
and compiler/build tool. So specifically mentioning it here makes
no sense and causes panic.

Signed-off-by: Sevki <s@sevki.org>