workspace: introduce the workspace package
this commit introduces the workspace package, removing a lot of
path related stuf from label. What I imagine I'll end up implementing
is git, http and local will be their own kinds of workspaces. This
will allow us to implement new ones in the future if necessary. An s3
workspace or a mercurial workspace for instance.

Signed-off-by: Sevki <s@sevki.org>